convert preprocessor to munger, autocomplete tests back on
