+++ src/main/scala/ammonite/kernel/ReplKernel.scala
@@ -36,7 +36,7 @@ object ReplKernel {
   }
 
   def complete(text: String, position: Int, interp: Interpreter) = {
-    interp.pressy.complete(text, position, Preprocessor.importBlock(interp.eval.frames.head.imports))
+    interp.pressy.complete(text, position, Munger.importBlock(interp.eval.frames.head.imports))
   }
 
 }
